+++ backend/like/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from .views import LikeCreateView
+from .views import LikeCreateView, LikeList
 from .views import LikeDestroyView
 from .views import LikeListView
 
@@ -8,4 +8,5 @@ urlpatterns = [
     path('cook_like/', LikeCreateView.as_view(), name="cook-like"),
     path('cook_unlike/<int:like_id>/', LikeDestroyView.as_view(), name="cook-dislike"),
     path('cook_likes/', LikeListView.as_view(), name="cook-list"),
+    path('list/', LikeList.as_view(), name="like-list")
 ]
