+++ backend/like/views.py
@@ -1,7 +1,17 @@
+'''buit-in'''
+from collections import Counter
+import random
+
+'''Third-party'''
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework import generics
+from rest_framework.views import APIView
 
+'''local'''
+from cook.models import CookIngredient, Cook
+from refrigerator.models import Refrigerator
+from user.models import User
 from .models import Like
 from .serializers import LikeSerializer
 from .serializers import LikeListSerializer
@@ -37,3 +47,45 @@ class LikeDestroyView(generics.DestroyAPIView):
 class LikeListView(generics.ListAPIView):
     serializer_class = LikeListSerializer
     queryset = Like.objects.all()
+
+
+class LikeList(APIView):
+    def get(self, request):
+
+        # user_number = abcd1234
+        user, _ = User.objects.get_or_create(user_number=request.META["HTTP_AUTHORIZATION"])
+        if Like.objects.filter(user=user.id).exists():
+            likes = Like.objects.filter(user=user.id)
+
+            result = list()
+            for like in likes:
+                cook = like.cook
+                recipes = [
+                    {
+                        "number": recipe.number,
+                        "detail": recipe.detail,
+                        "image_route": recipe.image_route
+                    } for recipe in cook.cook_recipe.all()
+                ]
+
+                ingredients = [
+                    {"name": ingredient.name} for ingredient in cook.cook_ingredient.all()
+                ]
+
+                result.append(
+                    {
+                        "id": cook.pk,
+                        "image_route": cook.image_route,
+                        "name": cook.name,
+                        "is_like": True,
+                        "ingredient": cook.ingredient.replace('주재료', '').split(','),
+                        "recipes": recipes,
+                        "ingredients": ingredients
+
+                    }
+                )
+
+            response = Response()
+            response.data = result
+            return response
+        return Response('찜목록이 없습니다.')
